TeamDataRepository code refactor
