+++ data/src/main/java/scoreapp/testtask/com/data/repository/TeamDataRepository.java
@@ -30,7 +30,7 @@ public class TeamDataRepository implements TeamRepository {
 
     @Override
     public Observable<List<Team>> getTeams() {
-        return sharedPreferences.makeGetObservable().map(sharedPreferencesEntityDataMapper::transform);
+        return sharedPreferences.makeGetObservable().map(this.sharedPreferencesEntityDataMapper::transform);
     }
 
     @Override
@@ -38,11 +38,11 @@ public class TeamDataRepository implements TeamRepository {
         SharedPreferencesEntity  sharedPreferencesEntity = this.sharedPreferences.get();
         List<Team> teamList = new ArrayList<>();
 
-        if (sharedPreferencesEntity!=null) {
+        if (sharedPreferencesEntity != null) {
             teamList = sharedPreferencesEntity.getListTeam();
         }
 
-        if(isUniqueTeam(team, teamList)) {
+        if(this.sharedPreferencesEntityDataMapper.isUniqueTeam(team, teamList)) {
             teamList.add(team);
             return this.sharedPreferences.makePutObservable(new SharedPreferencesEntity(teamList));
         } else {
@@ -54,87 +54,12 @@ public class TeamDataRepository implements TeamRepository {
     public Completable updateTeam(Team team) {
         List<Team>  teamList = this.sharedPreferences.get().getListTeam();
 
-        return this.sharedPreferences.makePutObservable(updateSharedPreferencesEntity(team, teamList));
+        return this.sharedPreferences.makePutObservable(
+                this.sharedPreferencesEntityDataMapper.transformForUpdate(team, teamList));
     }
 
     @Override
     public Completable clearSharedPreferences() {
         return this.sharedPreferences.makeClearObservable();
-    }
-
-    private SharedPreferencesEntity updateSharedPreferencesEntity(Team team, List<Team> listTeam) {
-        int teamIdToAddOppositeTeam = getTeamId(team,listTeam);
-        OppositeTeam newOppositeTeam = team.getOppositeTeam(0);
-
-        if (!isUniqueTeam(team, listTeam)) {
-            for(int teamId = 0; teamId < listTeam.size(); teamId++){
-                List<OppositeTeam> OppositeTeams = listTeam.get(teamId).getOppositeTeamList();
-
-                for(int OppositeTeam = 0; OppositeTeam < OppositeTeams.size(); OppositeTeam++ ) {
-
-                    if(!isUniqueTeam(newOppositeTeam, OppositeTeams) && teamId == teamIdToAddOppositeTeam){
-                        OppositeTeams.remove(getOppositeTeamId(newOppositeTeam, OppositeTeams));
-                        OppositeTeams.add(newOppositeTeam);
-                    } else if(teamId == teamIdToAddOppositeTeam) {
-                        listTeam.get(teamIdToAddOppositeTeam).addOppositeTeam(newOppositeTeam);
-                        break;
-                    }
-                }
-
-                if(OppositeTeams.size() == 0 && teamId == teamIdToAddOppositeTeam) {
-                    listTeam.get(teamIdToAddOppositeTeam).addOppositeTeam(newOppositeTeam);
-                }
-            }
-        }
-
-        return new SharedPreferencesEntity(listTeam);
-    }
-
-    private boolean isUniqueTeam(Team team, List<Team> listTeam) {
-        boolean isUniqueTeam = true;
-
-        for(int i=0; i<listTeam.size(); i++) {
-            if(listTeam.get(i).getTeamName().equals(team.getTeamName())) {
-                isUniqueTeam = false;
-            }
-        }
-
-        return isUniqueTeam;
-    }
-
-    private boolean isUniqueTeam(OppositeTeam team, List<OppositeTeam> listTeam) {
-        boolean isUniqueTeam = true;
-
-        for(int i=0; i<listTeam.size(); i++) {
-            if(listTeam.get(i).getTeamName().equals(team.getTeamName())) {
-                isUniqueTeam = false;
-            }
-        }
-
-        return isUniqueTeam;
-    }
-
-    private int getOppositeTeamId(OppositeTeam team, List<OppositeTeam> listTeam) {
-        int teamId = -1;
-
-        for(int i=0; i<listTeam.size(); i++) {
-            if(listTeam.get(i).getTeamName().equals(team.getTeamName())) {
-                teamId = i;
-            }
-        }
-
-        return teamId;
-    }
-
-    private int getTeamId(Team team, List<Team> listTeam) {
-        int teamId = -1;
-
-        for(int i=0; i<listTeam.size(); i++) {
-            if(listTeam.get(i).getTeamName().equals(team.getTeamName())) {
-                teamId = i;
-            }
-        }
-
-        return teamId;
     }
 }
